+++ gradle/libs.versions.toml
@@ -45,7 +45,6 @@ ktlintGradle = "11.6.1"
 paparazzi = "1.3.2"
 totp = "1.7.1"
 uiTestJunit4 = "1.5.4"
-material = "1.9.0"
 
 [bundles]
 androidx = ["androidx-ktx", "androidx-lifecycle", "androidx-activity-compose"]
@@ -126,7 +125,6 @@ strikt-mockk = { module = "io.strikt:strikt-mockk", version.ref = "striktMockk" 
 testlab-instr-lib = { module = "com.google.firebase:testlab-instr-lib", version.ref = "testlabInstrLib" }
 test-parameter-injector = { module = "com.google.testparameterinjector:test-parameter-injector", version.ref = "testParameterInjector" }
 totp = { module = "dev.samstevens.totp:totp", version.ref = "totp" }
-material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "androidGradlePlugin" }
